requirements.txt
discord.py
python-dotenv
os.python
platform
youtube_dl.py
time.py
asyncio.py
datetime.py

